--- src/main/java/com/example/fourchelin/domain/search/entity/PopularSearch.java
@@ -1,23 +1,0 @@
-package com.example.fourchelin.domain.search.entity;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDate;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class PopularSearch {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    @Column(nullable = false)
-    private String popularKeyword;
-    @Column(nullable = false)
-    private Long searchCount;
-    @Column(nullable = false)
-    private LocalDate trendDate;
-}
