--- src/main/java/com/example/fourchelin/domain/search/exception/SearchException.java
@@ -1,7 +1,0 @@
-package com.example.fourchelin.domain.search.exception;
-
-public class SearchException extends RuntimeException {
-    public SearchException(String message) {
-        super(message);
-    }
-}
