+++ src/main/java/com/example/fourchelin/domain/store/controller/StoreController.java
@@ -15,6 +15,18 @@ public class StoreController {
 
     private final StoreService storeService;
 
+    @GetMapping("/v1/stores")
+    public RspTemplate<StorePageResponse> searchStore(@RequestParam String keyword,
+                                                      @RequestParam(defaultValue = "1") int page,
+                                                      @RequestParam(defaultValue = "10") int size,
+                                                      @RequestParam(required = false) int star,
+                                                      @RequestParam(required = false) String status) {
+        // 필터링된 업체 리스트 반환
+        StorePageResponse res = searchService.searchStore(keyword, page, size, star,status);
+        return new RspTemplate<>(HttpStatus.OK, res);
+
+    }
+
     @GetMapping()
     public RspTemplate<StorePageResponse> findStores(@RequestParam(required = false) String keyword,
                                                      @RequestParam(required = false) String category,
