+++ src/main/java/com/example/fourchelin/domain/store/dto/response/StoreResponse.java
@@ -4,6 +4,8 @@ import com.example.fourchelin.domain.store.entity.Store;
 import com.example.fourchelin.domain.store.enums.StoreCategory;
 import com.example.fourchelin.domain.store.enums.StoreStatus;
 
+import java.time.LocalDateTime;
+
 public record StoreResponse(
         Long id,
         String storeName,
@@ -24,3 +26,28 @@ public record StoreResponse(
 
     }
 }
+//package com.example.fourchelin.domain.search.dto.response;
+//
+//import com.example.fourchelin.domain.store.entity.Store;
+//
+//import java.time.LocalDateTime;
+//
+//public record SearchResponse(
+//        Long id,
+//        String storeName,
+//        String address,
+//        int star,
+//        String status,
+//        LocalDateTime monitoringDate
+//) {
+//    public static SearchResponse fromStoreEntity(Store store) {
+//        return new SearchResponse(
+//                store.getId(),
+//                store.getStoreName(),
+//                store.getAddress(),
+//                store.getStar(),
+//                store.getStatus().name(), // enum 상태를 문자열로 변환
+//                store.getMonitoringDate()
+//        );
+//    }
+//}
